+++ prqlc/prqlc-ast/src/lib.rs
@@ -1,10 +1,9 @@
-pub mod error;
 pub mod expr;
-mod span;
+pub mod span;
 pub mod stmt;
 mod types;
 
 pub use expr::*;
-pub use span::Span;
+pub use span::*;
 pub use stmt::*;
 pub use types::*;